liked questions

+++ app/controllers/liked_questions_controller.rb
@@ -61,12 +61,17 @@ class LikedQuestionsController < ApplicationController
   # DELETE /liked_questions/1
   # DELETE /liked_questions/1.json
   def destroy
-    LikedQuestion.delete_all(["question_id=? AND user_id=?", params[:id], session[:user_id]])
-    question = Question.find(params[:id])
-    respond_to do |format|
-      format.html { redirect_to question_url(question) }
-      format.json { head :no_content }
+    if (Question.find(params[:question_id]).user_id != session[:user_id]) && (DislikedQuestion.where({question_id: params[:question_id], user_id: session[:user_id]}).empty?)
+      LikedQuestion.delete_all(["question_id=? AND user_id=?", params[:id], session[:user_id]])
+      DislikedQuestion.create(question_id: params[:question_id], user_id: session[:user_id])
+      question = Question.find(params[:id])
+      respond_to do |format|
+        format.html { redirect_to question_url(question) }
+        format.json { head :no_content }
+      end
+    else
     end
+
   end
 
   private
